Move CommandLine implementation to its own file
